english-reader-server/requirements.txt
fastapi
uvicorn
spacy
pdfplumber
python-docx
python-multipart
google-genai
python-dotenv
